cm_crowdin: combine the two sync scripts
cm_crowdin_sync_all.py is a simple merge of the two sync scripts.
It doesn't contain any optimisations. It should just be able to sync
properly with Crowdin (no interference from the other project) and
combine all translations for a project in a single Gerrit change.

PS5: Make the script executable

Change-Id: I361825d14757aa110a725b6644b9c21321210113